Mordernizing demo
WIP This is a first commit on the way from 2.5 to Lift 3.1.
At this point we have moved to 3.1 but some scala files is remed and awaits porting to 3.1.
The new layout is now responsive and uses Bootstrap v4 but the look is stil simular to the original.
ToDo's is plenty, see Notes.md for more info.